Added comments and some conditions for less bugs

+++ src/CtF.java
@@ -5,16 +5,20 @@ public class CtF
 
 	public static void main(String[] args) 
 	{
+		//I declared a scanner variable and longs just in case someone put in a big input
 		Scanner input = new Scanner(System.in);
-		int minutes, minutesRem, hours, days, years;
+		long minutes, minutesRem, hours, days, years;
 		
+		//Asks the user to enter the minutes in
 		System.out.println("Enter a number of minutes");
-		minutes = input.nextInt();
+		minutes = input.nextLong();
 
+		//Calculations for years, days, and remaining minutes to toss out
 		years = minutes / 525600;
 		minutesRem = minutes % 525600;
 		days = minutesRem / 1440;
 		
+		//Results!
 		System.out.println(minutes + " minutes is around " + years + " years and " + days + " days.");
 
 	}
